+++ Assets/UI/Popup/LosePopupUI.cs
@@ -1,3 +1,4 @@
+using MaskTransitions;
 using TMPro;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -21,11 +22,17 @@ public class LosePopupUI : MonoBehaviour
     void OnEnable()
     {
         _playAgainButton.onClick.AddListener(ReplayLevel);
+        _mainMenuButton.onClick.AddListener(OnMainMenuButtonClicked);
     }
+    private void OnMainMenuButtonClicked()
+    {
+        TransitionManager.Instance.LoadLevel(Others.MENU_SCENE);
+    }
+
     private void ReplayLevel()
     {
         GameManager.Instance.ResetInstance();
-        SceneManager.LoadScene(Others.GAME_SCENE);
+        TransitionManager.Instance.LoadLevel(Others.GAME_SCENE);
     }
     private void OnGameLose(string currentClock)
     {