+++ Assets/UI/Popup/WinPopupUI.cs
@@ -1,4 +1,5 @@
 using System;
+using MaskTransitions;
 using TMPro;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -22,13 +23,21 @@ public class WinPopupUI : MonoBehaviour
     void OnEnable()
     {
         _playAgainButton.onClick.AddListener(ReplayLevel);
+        _mainMenuButton.onClick.AddListener(OnMainMenuButtonClicked);
     }
+
+    private void OnMainMenuButtonClicked()
+    {
+        TransitionManager.Instance.LoadLevel(Others.MENU_SCENE);
+    }
+
     private void OnGameWin(string currentClock)
     {
         _winPopupTimer.text = currentClock;
     }
-    private void ReplayLevel() {
+    private void ReplayLevel()
+    {
         GameManager.Instance.ResetInstance();
-        SceneManager.LoadScene(Others.GAME_SCENE);
+        TransitionManager.Instance.LoadLevel(Others.GAME_SCENE);
     }
 }